refactor: move GlobalConfigPostProcessor to its own file
