+++ tests/testdata/test_addons/package_global_config_inputs_configuration_alerts/globalConfig.json
@@ -1108,7 +1108,7 @@
     "meta": {
         "name": "Splunk_TA_UCCExample",
         "restRoot": "splunk_ta_uccexample",
-        "version": "5.20.0R63ec1d0d",
+        "version": "5.20.0R1a58f15a",
         "displayName": "Splunk UCC test Add-on",
         "schemaVersion": "0.0.3"
     }
